+++ src/pages/namespace/_instances/Detail/index.tsx
